Fix sendmessages attachments

+++ SearchReactions.py
@@ -393,7 +393,7 @@ async def SendMessages(ctx: commands.Context):
                 if 'embed_description' in message:
                     embed = discord.Embed(description=message['embed_description'])
 
-                files = []
+                local_files = []
                 if 'attachments' in message:
                     attachments = json.loads(message['attachments'])
                     local_files = []
@@ -405,10 +405,12 @@ async def SendMessages(ctx: commands.Context):
                             os.remove(path)
                         download_image(file['url'], path)
                         local_files.append(discord.File(path, filename=file['filename']))
-                    for i in range(len(attachments)):
-                        os.remove(f'{tmp_folder}/{i}')
 
-                await user.send(message['text'], files=files, embed=embed, view=view)
+                await user.send(message['text'], files=local_files, embed=embed, view=view)
+
+                if 'attachments' in message:
+                    for i in range(len(local_files)):
+                        os.remove(f'{tmp_folder}/{i}')
 
                 message["message_sent"] = 1
                 db_items("Messages").upsert_item(body=message)
